refactor: update tsconfig
