+++ src/constants/constants.ts
@@ -1,0 +1,1 @@
+export const URL_SEPARATOR = '/';
